III-3137 Move Offer history project methods to trait
